requirements.txt
streamlit
pytesseract
opencv-python
Pillow
PyMuPDF
transformers
torch
spacy 